♻️  Add version field and configure injection of extension version in GithubAction class
This commit adds a new version field to GithubAction class and configures the injection of extension version from the pom.xml. The field's access level is set to package-private to avoid excessive reflection usage. This will make it possible for the class to read the version of the extension and will allow for more efficient reflection usage.