✨ Regenerate lockfile if workflow file is changed
Add support for specifying a custom name for the workflow file in the input. This enables users to use any name they want for the file, rather than being restricted to using the default name.

This commit also updates the step for changed files to include the workflow file, when specified in the input.